Write texture coordinates into the converted obj files and create material files for them
Currently only objects with a single texture are supported. Texture coordinates are read from the polygon_mappings where texture normals are read from as well. While the normals are stored in v3 as (x, y, z) the texture coordinates u,v are stored in v1 and v2 where v1 contains all us and v2 all vs of the vertices.