Add option to copy texture files next to the converted files
If a texture search path is given, the texture files referenced in a bgf
will be copied next to the converted obj/mtl files so that 3D viewers
can find them.
This requires unpacking the textures.bin (a zip archive) first and could
be extended to search the files from the bin directly.