Add incomplete support for multi model bgf files
Also experimentally use the game_objects data instead of the
vertex/polygon_mapping data as input. Here every model has its own list
instead of a common merged list as in the mapping object. This makes it
easer to work with and it seems to work for the simple models.
More complex models still have distortions and glitches. Whether this
stems from misinterpretation of the data/bugs or the usage of the other
data source has to be investigated next.