Add very basic implementation of animation decoding
Currently not all animations are supported (there are some unhandled special cases in the file format) and the animation files are not converted yet, just displayed using matplotlib. This plot is pretty basic and doesn't actually use the real model with textures and faces, instead it only plots the vertices.